HSEARCH-3250 Library showcase: introduce an AbstractEntity class to simplify entity code
