+++ integrationtest/showcase/library/src/main/java/org/hibernate/search/v6poc/integrationtest/showcase/library/model/AbstractEntity.java
@@ -1,0 +1,48 @@
+/*
+ * Hibernate Search, full-text search for your domain model
+ *
+ * License: GNU Lesser General Public License (LGPL), version 2.1 or later
+ * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
+ */
+package org.hibernate.search.v6poc.integrationtest.showcase.library.model;
+
+import java.util.Objects;
+import javax.persistence.MappedSuperclass;
+
+import org.hibernate.Hibernate;
+
+@MappedSuperclass
+public abstract class AbstractEntity<ID> {
+
+	@Override
+	public boolean equals(Object o) {
+		if ( this == o ) {
+			return true;
+		}
+		if ( o == null || getClass() != Hibernate.getClass( o ) ) {
+			return false;
+		}
+		AbstractEntity<?> other = (AbstractEntity<?>) o;
+		return Objects.equals( getId(), other.getId() );
+	}
+
+	@Override
+	public int hashCode() {
+		return Objects.hash( getClass() );
+	}
+
+	@Override
+	public String toString() {
+		return new StringBuilder( getClass().getSimpleName() )
+				.append( "[" )
+				.append( getId() )
+				.append( "," )
+				.append( getDescriptionForToString() )
+				.append( "]" )
+				.toString();
+	}
+
+	public abstract ID getId();
+
+	protected abstract String getDescriptionForToString();
+}
