+++ integrationtest/showcase/library/src/main/java/org/hibernate/search/v6poc/integrationtest/showcase/library/model/Document.java
@@ -8,7 +8,6 @@ package org.hibernate.search.v6poc.integrationtest.showcase.library.model;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
 import javax.persistence.Basic;
 import javax.persistence.Entity;
 import javax.persistence.Id;
@@ -16,10 +15,8 @@ import javax.persistence.Inheritance;
 import javax.persistence.InheritanceType;
 import javax.persistence.OneToMany;
 
-import org.hibernate.Hibernate;
 import org.hibernate.search.v6poc.backend.document.model.dsl.ObjectFieldStorage;
 import org.hibernate.search.v6poc.backend.document.model.dsl.Sortable;
-import org.hibernate.search.v6poc.entity.pojo.mapping.definition.annotation.DocumentId;
 import org.hibernate.search.v6poc.entity.pojo.mapping.definition.annotation.Field;
 import org.hibernate.search.v6poc.entity.pojo.mapping.definition.annotation.IndexedEmbedded;
 import org.hibernate.search.v6poc.integrationtest.showcase.library.bridge.annotation.MultiKeywordStringBridge;
@@ -31,10 +28,9 @@ import org.hibernate.search.v6poc.integrationtest.showcase.library.bridge.annota
  */
 @Entity
 @Inheritance(strategy = InheritanceType.TABLE_PER_CLASS)
-public abstract class Document<C extends DocumentCopy<?>> {
+public abstract class Document<C extends DocumentCopy<?>> extends AbstractEntity<Integer> {
 
 	@Id
-	@DocumentId
 	private Integer id;
 
 	@Basic
@@ -58,31 +54,6 @@ public abstract class Document<C extends DocumentCopy<?>> {
 	private List<C> copies = new ArrayList<>();
 
 	@Override
-	public boolean equals(Object o) {
-		if ( this == o ) {
-			return true;
-		}
-		if ( o == null || getClass() != Hibernate.getClass( o ) ) {
-			return false;
-		}
-		Document<?> document = (Document<?>) o;
-		return Objects.equals( id, document.getId() );
-	}
-
-	@Override
-	public int hashCode() {
-		return Objects.hash( getClass() );
-	}
-
-	@Override
-	public String toString() {
-		return new StringBuilder( getClass().getSimpleName() )
-				.append( "[" )
-				.append( id )
-				.append( "]" )
-				.toString();
-	}
-
 	public Integer getId() {
 		return id;
 	}
@@ -123,4 +94,8 @@ public abstract class Document<C extends DocumentCopy<?>> {
 		this.copies = copies;
 	}
 
+	@Override
+	protected String getDescriptionForToString() {
+		return getTitle();
+	}
 }
